backend/requirements.txt
# WebSocket and HTTP
websockets==15.0.1
aiohttp==3.12.14

# Database
pymongo==4.13.2

# Environment and configuration
python-dotenv==1.1.1

# Async utilities
asyncio-mqtt==0.16.1

# Logging and utilities
colorlog==6.8.0

# Optional: For enhanced logging
rich==13.7.0 